lich su don hang

+++ Entity/EntityFramework/MERCHANT.cs
@@ -36,7 +36,7 @@ namespace Entity.EntityFramework
         public Nullable<int> SOLANBIKHOA { get; set; }
         public Nullable<System.DateTime> NGAYSINH { get; set; }
         public string SDT { get; set; }
-        public Nullable<int> RATING { get; set; }
+        public Nullable<double> RATING { get; set; }
         public Nullable<int> SOTINHIENTAI { get; set; }
     
         [System.Diagnostics.CodeAnalysis.SuppressMessage("Microsoft.Usage", "CA2227:CollectionPropertiesShouldBeReadOnly")]
